% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wsj.R
\name{theme_wsj}
\alias{theme_wsj}
\title{Make timeseries graphs look like the the Wall Street Journal.}
\usage{
theme_wsj()
}
\description{
Make timeseries graphs look like the the Wall Street Journal.
}
\examples{
library(ggplot2)
`\%>\%` <- magrittr::`\%>\%`

  filter(variable \%in\% c("psavert", "uempmed")) \%>\%
  ggplot(aes(date, value, color = variable)) +
  geom_line() +
  scale_y_continuous(
      labels = label_wsj(suffix = " M")
  ) +
  scale_color_discrete(
      labels = c("Series 1", "Series 2")
  ) +
  theme_wsj() +
  labs(
    title = "Some Economics Plot",
    caption = "Source: Not well documented.",
    y = ""
  )

}
